crate apis

+++ app/db/schema/reviews.py
@@ -3,11 +3,11 @@ from typing import Optional
 
 class PostReviews(BaseModel):
     book_id: int
-    ratings: int
+    ratings: float
     review: str
     reviewer_name: str
 
 class UpdateReviews(BaseModel):
     review_id: int
-    ratings: Optional[int] = None
+    ratings: Optional[float] = None
     review: Optional[str] = None
